Refactor exception to separate file
